Migrate switch from /docs/components/switch to /components/switch
Move all examples (Consent, Form, Notification Preferences) from the
old documentation page to the new reference page, remove the old page,
and update all navigation links and E2E tests.
