+++ site/ui/components/command-palette.tsx
@@ -31,7 +31,7 @@ const componentItems = [
   { id: 'dropdown-menu', title: 'Dropdown Menu', href: '/docs/components/dropdown-menu', category: 'Components' },
   { id: 'input', title: 'Input', href: '/components/input', category: 'Components' },
   { id: 'select', title: 'Select', href: '/components/select', category: 'Components' },
-  { id: 'switch', title: 'Switch', href: '/docs/components/switch', category: 'Components' },
+  { id: 'switch', title: 'Switch', href: '/components/switch', category: 'Components' },
   { id: 'tabs', title: 'Tabs', href: '/docs/components/tabs', category: 'Components' },
   { id: 'toast', title: 'Toast', href: '/docs/components/toast', category: 'Components' },
   { id: 'tooltip', title: 'Tooltip', href: '/docs/components/tooltip', category: 'Components' },
